Revert "Choose to create or link storefront on `init`"
This reverts commit 4764eacf7789d032d2e792613ffbfb49af672752.
