+++ packages/cli/src/commands/hydrogen/link.test.ts
@@ -54,9 +54,6 @@ describe('link', () => {
     storefront: undefined,
   };
 
-  const expectedStorefrontName = 'New Storefront';
-  const expectedJobId = 'gid://shopify/Job/1';
-
   beforeEach(async () => {
     vi.mocked(login).mockResolvedValue({
       session: ADMIN_SESSION,
@@ -72,15 +69,6 @@ describe('link', () => {
     ]);
 
     vi.mocked(renderSelectPrompt).mockResolvedValue(FULL_SHOPIFY_CONFIG.shop);
-
-    vi.mocked(createStorefront).mockResolvedValue({
-      storefront: {
-        id: 'gid://shopify/HydrogenStorefront/1',
-        title: expectedStorefrontName,
-        productionUrl: 'https://example.com',
-      },
-      jobId: expectedJobId,
-    });
   });
 
   afterEach(() => {
@@ -140,10 +128,20 @@ describe('link', () => {
   });
 
   describe('when you want to link a new Hydrogen storefront', () => {
-
+    const expectedStorefrontName = 'New Storefront';
+    const expectedJobId = 'gid://shopify/Job/1';
 
     beforeEach(async () => {
       vi.mocked(renderSelectPrompt).mockResolvedValue(null);
+
+      vi.mocked(createStorefront).mockResolvedValue({
+        storefront: {
+          id: 'gid://shopify/HydrogenStorefront/1',
+          title: expectedStorefrontName,
+          productionUrl: 'https://example.com',
+        },
+        jobId: expectedJobId,
+      });
     });
 
     it('chooses to create a new storefront given the directory path', async () => {
